pyproject.toml
[tool.poetry]
name = "funnyboard"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^2.0.1"
requests = "^2.28.1"
pydantic = "^1.10.2"
python-twitter = "^3.5"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
